+++ packages/verite/test/lib/issuer/credential-fulfillment.test.ts
@@ -173,7 +173,7 @@ describe("buildAndSignFulfillment", () => {
     // The fulfillment looks like this:
     expect(fulfillment).toMatchObject({
       "@context": ["https://www.w3.org/2018/credentials/v1"],
-      type: ["VerifiablePresentation", "CredentialFulfillment"],
+      type: ["VerifiablePresentation", "CredentialResponse"],
       holder: issuer.did,
       credential_fulfillment: {
         // id: "5f22f1ea-0441-4041-916b-2504a2a4075c",
@@ -271,7 +271,7 @@ describe("buildAndSignFulfillment", () => {
     const fulfillment = await decodeVerifiablePresentation(encodedFulfillment)
     expect(fulfillment).toMatchObject({
       "@context": ["https://www.w3.org/2018/credentials/v1"],
-      type: ["VerifiablePresentation", "CredentialFulfillment"],
+      type: ["VerifiablePresentation", "CredentialResponse"],
       holder: issuer.did,
       credential_fulfillment: {
         manifest_id: "HybridManifest",
