+++ packages/verite/types/CredentialResponse.ts
