+++ packages/verite/types/index.ts
@@ -3,7 +3,7 @@ export * from "./ChallengeTokenUrl"
 export * from "./ClaimFormatDesignation"
 export * from "./CredentialIssuer"
 export * from "./CredentialApplication"
-export * from "./CredentialFulfillment"
+export * from "./CredentialResponse"
 export * from "./CredentialManifest"
 export * from "./CredentialOffer"
 export * from "./DataDisplay"
